+++ messages/en/pages.php
@@ -49,6 +49,8 @@ return [
     'settings_delete'  => 'Delete page',
     'create_btn'       => 'Create',
     'menu_add_btn'     => 'Create a new content page in the menu',
+    'fullscr_select'   => 'Content pages',
+    'fullscr_split'    => 'Split screen',
 
     'documents_title' => 'Documents',
     'documents_intro_text' => 'On this page, documents for this consultation are published.',